Separada la logida de JavaScript en un hook personalizado: useVideoLogic.js, ademas se configuro el archivo peticionesVideo.js porque daba falla al realizar las solicitudes post y patch
